ide: CLANG now default method in win32
git-svn-id: svn://ultimatepp.org/upp/trunk@14378 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/ide/Debuggers/PrettyStd.cpp
@@ -238,7 +238,7 @@ void Pdb::PrettyStdDeque(Pdb::Val val, const Vector<String>& tparam, int64 from,
 	}
 	
 	for(int i = 0; i < count; i++) {
-		int q = i + from + start;
+		int64 q = i + from + start;
 		p.data_ptr.Add(PeekPtr(map + (q / block_size) * (win64 ? 8 : 4)) + q % block_size * sz);
 	}
 }
